Prepare for Reality Upgrade addition.
Learn your ABCs, Earth.
